fix: refactor & simplify typing animation
