+++ src/domain/props/TypingTextProps.tsx
@@ -1,8 +1,7 @@
 // TypingTextProps is a type that defines the props for the TypingText component
 export default interface TypingTextProps {
   text: string;
-  id: string;
-  speed?: number;
+  textID: string;
   isUserMessageSent?: boolean;
   chatIndex?: number;
   onTypingStart?: (chatIndex: number, isTyping: boolean) => void;
